When VEROBSE flag is set greater than 0, print traces in vm_execute_all
For each cycle, print the cycle and any changes.  We track changes on
the stack by remembering the previous stack pointer.  For registers, I
remember the previous array of registers and do a byte level compare
of the current registers and the remembered ones.

Produces pretty log messages and an easy way to track execution.