Defined a base testing library header
This defines useful macros, in particular to print information and to
define tests and test suites.

The idea is a suite will be a set of test functions, where a test
function will fail <=> an assert is called.  This allows us to stop
testing immediately upon a test failing and produce viable output.