bytecode_t is a typedef over darr_t
Simpler to use than a custom structure for users of C.

For FFI users the layout is no different to a custom structure anyway.
As long as we don't use ANY memory management stuff as part of the
darr (i.e. ensure_capacity, append, etc), same as considering the
darr_t as not really owning the pointer itself, we should be fine.